Catch ProcessLookupError in signal forwarding
When testing using papermill in Python 3.8 (only), we see this exception
(for signal 17 SIGCHLD) which we try to forward.

We do a double fix, don't try to forward this signal, because it's for
us and not for the child (kernel).

And ignore the exception because if it happens, the process has already
exited anyway.