Use jupyter-client to spawn kernel process
Use jupyter-client to spawn the kernel process and adopt its method for
signaling. This matters for windows (and does not seem to matter for
Linux or Macos).

There is a change on posix platforms, we now use process group id and
killpg to send SIGINT to the process group (as done by jupyter).
jupyter-client now spawns the kernel process in a separate process
group.

We include the file parentpoller.py from ipykernel licensed under a
BSD 3-clause license which is compatible with the project.