Implement the main run fallback
This is only used if we are not launched using jupyter-client, which in
practice means from VSCodium or VS Code as far as I know.

The main use case for pyproject_local_kernel is to use with jupyterlab,
so the scenario implemented in this commit is less important - for now
it is logged as unsupported.